requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==41.0.4
Django==4.2.6
djangorestframework==3.14.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
